options for upgrade install

+++ install.sh
@@ -15,7 +15,6 @@ else
 fi
 sleep 1
 [ -f /usr/local/sbin/brubeck ] && mv /usr/local/sbin/brubeck /usr/local/sbin/brubeck.old
-[ -f /etc/brubeck/config.json ] && mv /etc/brubeck/config.json /etc/brubeck/config.json.old
 [ -f /etc/init.d/brubeck ] && mv /etc/init.d/brubeck /etc/init.d/brubeck.old
 #
 # install the binary
@@ -29,32 +28,67 @@ chmod +x /usr/local/sbin/brubeck
 # install the init file
 echo "$STEP/$TOTAL Installing init script ..."
 STEP=$((STEP+1))
-if [ -d /etc/init.d ] ; then
+VALUE=n
+read -p "Install init file [(y/n)n]: " RVALUE
+[ ! -z "${RVALUE}" ] && VALUE="${RVALUE}"
+if [ $VALUE == y ] ; then
+    if [ -d /etc/init.d ] ; then
 	cp brubeck_init.d /etc/init.d/brubeck
 	chown root.root /etc/init.d/brubeck
 	chmod +x /etc/init.d/brubeck
 	if [ `which systemctl` ] ; then
-	systemctl daemon-reload > /dev/null 2>&1
-	if [ `which chkconfig` ] ; then
-	    chkconfig --add brubeck
-	else
-	    if [ `which update-rc.d` ] ; then
-		update-rc.d brubeck defaults
+	    systemctl daemon-reload > /dev/null 2>&1
+	    if [ `which chkconfig` ] ; then
+		chkconfig --add brubeck
 	    else
-		systemctl enable brubeck
+		if [ `which update-rc.d` ] ; then
+		    update-rc.d brubeck defaults
+		else
+		    systemctl enable brubeck
+		fi
 	    fi
-	fi
 	else
-	echo "No systemctl found.  Service not configured.  Init script installed."
+	    echo "No systemctl found.  Service not configured.  Init script installed."
 	fi
+    fi
 fi
+
+
+function set()
+{
+   MUNGE=$1
+   KEY=$2
+   VALUE=$3
+   FILE=$4
+
+   sed -i -e "s/${MUNGE}/            \"${KEY}\" : \"${VALUE}\"\,/g" $FILE
+}
+
+function getset()
+{
+   MUNGE=$1
+   KEY=$2
+   VALUE=$3
+   PROMPT=$4
+   FILE=$5
+
+   read -p "${PROMPT} [${VALUE}]: " RVALUE
+   [ ! -z "${RVALUE}" ] && VALUE="${RVALUE}"
+
+   set "${MUNGE}" "${KEY}" "${VALUE}" "${FILE}"
+}
 
 echo "$STEP/$TOTAL Configuring ..."
 STEP=$((STEP+1))
 echo ""
-CONFIG_WORK="/tmp/config.json"
-rm -f "${CONFIG_WORK}"
-cat <<EOF >> "${CONFIG_WORK}"
+VALUE=n
+read -p "Reconfigure config file [(y/n)n]: " RVALUE
+[ ! -z "${RVALUE}" ] && VALUE="${RVALUE}"
+if [ $VALUE == y ] ; then
+    [ -f /etc/brubeck/config.json ] && mv /etc/brubeck/config.json /etc/brubeck/config.json.old
+    CONFIG_WORK="/tmp/config.json"
+    rm -f "${CONFIG_WORK}"
+    cat <<EOF >> "${CONFIG_WORK}"
 {
     "sharding" : false,
 SERVER_NAME
@@ -84,44 +118,20 @@ BRUBECK_PORT
 }
 EOF
 
-function set()
-{
-   MUNGE=$1
-   KEY=$2
-   VALUE=$3
-   FILE=$4
+    # get the vars
+    # server_name
+    HOSTNAME=`uname -n`
+    getset "SERVER_NAME" "server_name" "brubeck-${HOSTNAME}" "Enter the server name, used for server metrics" "${CONFIG_WORK}"
+    getset "CARBON_SERVER" "address" "bubble.bottorrent.net" "Enter a graphite/carbon server ip name or number" "${CONFIG_WORK}"
+    getset "BRUBECK_PORT" "port" "8125" "Enter the listening port for brubeck, typically 8125." "${CONFIG_WORK}"
 
-   sed -i -e "s/${MUNGE}/            \"${KEY}\" : \"${VALUE}\"\,/g" $FILE
-}
-
-function getset()
-{
-   MUNGE=$1
-   KEY=$2
-   VALUE=$3
-   PROMPT=$4
-   FILE=$5
-
-   read -p "${PROMPT} [${VALUE}]: " RVALUE
-   [ ! -z "${RVALUE}" ] && VALUE="${RVALUE}"
-
-   set "${MUNGE}" "${KEY}" "${VALUE}" "${FILE}"
-}
-
-# get the vars
-# server_name
-HOSTNAME=`uname -n`
-getset "SERVER_NAME" "server_name" "brubeck-${HOSTNAME}" "Enter the server name, used for server metrics" "${CONFIG_WORK}"
-getset "CARBON_SERVER" "address" "bubble.bottorrent.net" "Enter a graphite/carbon server ip name or number" "${CONFIG_WORK}"
-getset "BRUBECK_PORT" "port" "8125" "Enter the listening port for brubeck, typically 8125." "${CONFIG_WORK}"
-
-read -p "Add datadog configuration [(y/n)n]: " RVALUE
-case $RVALUE in
+    read -p "Add datadog configuration [(y/n)n]: " RVALUE
+    case $RVALUE in
 	Y|y)
-# the datadog blob
-DATADOG_CONFIG_WORK="/tmp/datadog.config.json"
-rm -f "${DATADOG_CONFIG_WORK}"
-cat <<EOF >> "${DATADOG_CONFIG_WORK}"
+	    # the datadog blob
+	    DATADOG_CONFIG_WORK="/tmp/datadog.config.json"
+	    rm -f "${DATADOG_CONFIG_WORK}"
+	    cat <<EOF >> "${DATADOG_CONFIG_WORK}"
 	},
 	{
 	    "type" : "datadog",
@@ -131,19 +141,20 @@ DATADOG_PORT
 DATADOG_FILTER
 	    "expire" : 1
 EOF
-	getset "DATADOG_SERVER" "address" "localhost" "Enter a datadog server, typically localhost" "${DATADOG_CONFIG_WORK}"
-	getset "DATADOG_PORT" "port" "9125" "Enter the port of the datadog server, typically 9125 as we run brubeck on 8125" "${DATADOG_CONFIG_WORK}"
-	getset "DATADOG_FILTER" "filter" ".*" "Enter a regex.  Matches will be sent to datadog.  Defaults to all." "${DATADOG_CONFIG_WORK}"
-   	sed -i -e "/\"pickle\" : false/r${DATADOG_CONFIG_WORK}" "${CONFIG_WORK}"
-	;;	
+	    getset "DATADOG_SERVER" "address" "localhost" "Enter a datadog server, typically localhost" "${DATADOG_CONFIG_WORK}"
+	    getset "DATADOG_PORT" "port" "9125" "Enter the port of the datadog server, typically 9125 as we run brubeck on 8125" "${DATADOG_CONFIG_WORK}"
+	    getset "DATADOG_FILTER" "filter" ".*" "Enter a regex.  Matches will be sent to datadog.  Defaults to all." "${DATADOG_CONFIG_WORK}"
+   	    sed -i -e "/\"pickle\" : false/r${DATADOG_CONFIG_WORK}" "${CONFIG_WORK}"
+	    ;;	
 
 	*)
-	;;
-esac
+	    ;;
+    esac
 
-[ ! -d /etc/brubeck ] && mkdir -p /etc/brubeck
-cp "${CONFIG_WORK}" /etc/brubeck/config.json
-chown -R root.root /etc/brubeck/config.json
+    [ ! -d /etc/brubeck ] && mkdir -p /etc/brubeck
+    cp "${CONFIG_WORK}" /etc/brubeck/config.json
+    chown -R root.root /etc/brubeck/config.json
+fi
 
 echo ""
 echo "$STEP/$TOTAL Restarting ..."
